Close M1.1: the catalog, and the hull stops being a loose byte
ADR-006's tables, referred to by identity and never by name. Five hulls, two
drives, three weapon components and four module ones, `constexpr` in GameCore
because R16 keeps simulation data where the two sides cannot disagree about it.

THE PLAN SAYS FOUR HULLS AND GameDesign.md SECTION 6 SAYS FIVE. The design is the
source and it is right: the station is a hull and so is a module frame, which is
what makes section 5 possible without a second kind of thing in the simulation
(ADR-015). Five are here. The plan's line is a miscount rather than a decision.

THE PLACEHOLDER IS GONE, WHICH WAS THE REAL WORK. `Entity.h` carried
`using HullId = std::uint8_t` with a note saying "at M0 there is no catalog to
index into and this is a number that rides along". There is one now. The
identity is the catalog's, the enumeration is uint8-backed so the state hash
folds the same byte it always did, and seven call sites stopped passing loose
integers.

AND THAT TURNED UP A NUMBER THAT WAS NEVER A HULL. The scripted determinism run
built three entities with hulls 2, 9 and 1 -- and there is no hull 9. It did not
matter while the field was a byte; it cannot be written at all now. Two of the
three are unchanged and the 9 became `ModuleFrame`, so the pinned hash moved from
0xd2a4d77a1900bf40 to 0xd2a4e47a1900d557 for exactly that reason and no other.

**MOVING THAT LITERAL IS THE ONE THING IN THIS COMMIT THAT DESERVED CARE**, so it
was verified rather than assumed: the same number comes out of Debug|x64,
Release|x64, Debug|ARM64 and Release|ARM64. That is the cross-platform assertion
the literal exists for and this is the first time it has been checked on ARM64
hardware rather than trusted. The reason it moved is written above it, along with
what a future disagreement would mean -- deliberate, or a desynchronisation.

WHAT THE CATALOG DELIBERATELY DOES NOT CARRY:

  MASS AND THRUST. Section 6 states a hull's mass as "low", "medium" or "high"
  and a drive as "balanced thrust, cheap" against "more thrust for more mass and
  more cost" -- relations, not figures -- while fixing the two outcomes they must
  produce: a Miner at 100 u/s and a Fighter at 140. Many sets of numbers satisfy
  that and they disagree about every hull the MVP does not build. That is a
  question for the register and M1.2 is where it is needed, which is the step
  that writes the derivation.

  SIZE IN WORLD UNITS. Q37, still open, and R24 wants the catalog to name it so a
  script can compare the figure against the mesh extent rather than a reader
  trusting both.

  ResearchStationL1. Section 6 gives it a dash for a cost and says it is designed
  at M4 when there is research for it to do; Q30 is the argument. The section's
  own count agrees -- "four module components" against five rows in its table.

SizeClass is `Light`/`Medium`/`Heavy` and cannot be otherwise: `<windows.h>`
defines `small` as a macro and the preprocessor rewrites it before the compiler
sees it. The mapping onto section 7's columns is a table in the header, because
the design's prose and this code will not match and a reader would reasonably
wonder whether that is a bug.

380 tests, four configuration pairs clean, gates clean, format clean.

+++ GameLogic/Host.cpp
@@ -67,7 +67,7 @@ Snapshot BuildSnapshot(const World& _world, const CommandIntake& _intake, std::u
                                              // degrees, which is a rendering quantity (TechnicalDesign.md section 4).
                                              .heading = static_cast<std::uint8_t>(entity.heading >> 8),
                                              .hullPercentRemaining = 100,
-                                             .designIdentity = entity.hull,
+                                             .designIdentity = static_cast<std::uint8_t>(entity.hull),
                                              .flags = static_cast<std::uint8_t>((entity.owner & FLAGS_TEAM_MASK) << FLAGS_TEAM_SHIFT)});
   }
